Solve: [Bug] Guest feed #537
There are several bugs in the guest feed:

Expected: We should be able to scroll down and we should get to the sign in page only by clicking on something. Currently: On mobile, when we scroll down we are automatically redirected to the sign in.
Add a "fake top bar" (home/squad/sign-in button: (see design), clicking on this top bar redirect to sign-in
Add the sign in button: should redirect to sign in